requirements.txt
aiohttp==3.9.1
pyyaml==6.0.1
websockets==12.0
python-dotenv==1.0.0
rich==13.7.0
requests==2.31.0
httpx==0.27.0 